# synapse.ai---lead-radar-&-code-defense/backend/tests/snapshots/openapi.json
{
 "components": {
  "schemas": {
...
   "ErrorResponse": {
    "description": "Every non-2xx response has this shape.",
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "detail": {
      "title": "Detail",
      "type": "string"
     },
     "errors": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/FieldError"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errors"
     }
    },
    "required": [
     "detail",
     "code"
    ],
    "title": "ErrorResponse",
    "type": "object"
   },
   "FieldError": {
    "properties": {
     "field": {
      "title": "Field",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "field",
     "message"
    ],
    "title": "FieldError",
    "type": "object"
   },
...
   "TranscriptTurn": {
    "properties": {
     "audioUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Audiourl"
     },
     "speaker": {
      "enum": [
       "ai_interviewer",
       "candidate"
      ],
      "title": "Speaker",
      "type": "string"
     },
     "technicalAssessmentNote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Technicalassessmentnote"
     },
     "text": {
      "title": "Text",
      "type": "string"
     },
     "timestamp": {
      "title": "Timestamp",
      "type": "string"
     }
    },
    "required": [
     "speaker",
     "text",
     "timestamp"
    ],
    "title": "TranscriptTurn",
    "type": "object"
   },
   "VoiceSessionInitRequest": {
    "properties": {
     "candidateName": {
      "anyOf": [
       {
        "maxLength": 255,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidatename"
     },
     "focusTopic": {
      "anyOf": [
       {
        "maxLength": 300,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Focustopic"
     },
     "repoId": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repoid"
     }
    },
    "title": "VoiceSessionInitRequest",
    "type": "object"
   },
   "VoiceSessionResponse": {
    "properties": {
     "architectureDefenseScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Architecturedefensescore"
     },
     "candidateName": {
      "title": "Candidatename",
      "type": "string"
     },
     "clarityArticulationScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clarityarticulationscore"
     },
     "completedAt": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completedat"
     },
     "defenseVerdict": {
      "title": "Defenseverdict",
      "type": "string"
     },
     "durationSeconds": {
      "title": "Durationseconds",
      "type": "integer"
     },
     "finalCompositeScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finalcompositescore"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "keyTakeaways": {
      "items": {
       "type": "string"
      },
      "title": "Keytakeaways",
      "type": "array"
     },
     "repoId": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repoid"
     },
     "sessionStatus": {
      "title": "Sessionstatus",
      "type": "string"
     },
     "startedAt": {
      "format": "date-time",
      "title": "Startedat",
      "type": "string"
     },
     "technicalDepthScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Technicaldepthscore"
     },
     "transcriptTurns": {
      "items": {
       "$ref": "#/components/schemas/TranscriptTurn"
      },
      "title": "Transcriptturns",
      "type": "array"
     }
    },
    "required": [
     "id",
     "candidateName",
     "sessionStatus",
     "startedAt",
     "durationSeconds",
     "defenseVerdict",
     "transcriptTurns",
     "keyTakeaways"
    ],
    "title": "VoiceSessionResponse",
    "type": "object"
   },
   "VoiceTurnInput": {
    "properties": {
     "candidateSpeechText": {
      "maxLength": 8000,
      "minLength": 1,
      "title": "Candidatespeechtext",
      "type": "string"
     },
     "sessionId": {
      "format": "uuid",
      "title": "Sessionid",
      "type": "string"
     }
    },
    "required": [
     "sessionId",
     "candidateSpeechText"
    ],
    "title": "VoiceTurnInput",
    "type": "object"
   },
   "VoiceTurnOutput": {
    "properties": {
     "aiStatus": {
      "description": "'unavailable' means no interviewer reply could be generated; the answer was still recorded",
      "enum": [
       "generated",
       "unavailable"
      ],
      "title": "Aistatus",
      "type": "string"
     },
     "currentCompositeScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "description": "Null: answers are not auto-scored",
      "title": "Currentcompositescore"
     },
     "currentVerdict": {
      "description": "Always a review state such as needs_human_review; never a hiring decision",
      "title": "Currentverdict",
      "type": "string"
     },
     "defenseScoreDelta": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "description": "Null: answers are not auto-scored",
      "title": "Defensescoredelta"
     },
     "interviewerReplyText": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interviewerreplytext"
     },
     "sessionId": {
      "format": "uuid",
      "title": "Sessionid",
      "type": "string"
     },
     "turnNumber": {
      "title": "Turnnumber",
      "type": "integer"
     }
    },
    "required": [
     "sessionId",
     "turnNumber",
     "aiStatus",
     "currentVerdict"
    ],
    "title": "VoiceTurnOutput",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "Synapse.AI",
  "version": "3.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/assessment/voice/init": {
   "post": {
    "operationId": "initialize_voice_session_api_v1_assessment_voice_init_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VoiceSessionInitRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VoiceSessionResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Start an interview session (returns the opening question and the session id to use)",
    "tags": [
     "Voice Defense"
    ]
   }
  },
...
  "/api/v1/assessment/voice/turn": {
   "post": {
    "operationId": "process_voice_turn_api_v1_assessment_voice_turn_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VoiceTurnInput"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VoiceTurnOutput"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Record a candidate answer and get the interviewer's follow-up (never a score or verdict)",
    "tags": [
     "Voice Defense"
    ]
   }
  },
  "/api/v1/assessment/voice/{session_id}": {
   "get": {
    "operationId": "get_session_api_v1_assessment_voice__session_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "session_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Session Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VoiceSessionResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "One session with its transcript",
    "tags": [
     "Voice Defense"
    ]
   }
  },
...
 }
}